moved regex processing
