+++ modules/openapi-generator/src/main/java/org/openapitools/codegen/languages/CSharpClientCodegen.java
@@ -37,7 +37,6 @@ import java.io.File;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
-import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import java.util.ArrayList;
 import java.util.Comparator;
@@ -622,58 +621,11 @@ public class CSharpClientCodegen extends AbstractCSharpCodegen {
 
     @Override
     public void postProcessParameter(CodegenParameter parameter) {
-        postProcessPattern(parameter.pattern, parameter.vendorExtensions);
+        super.postProcessParameter(parameter);
         postProcessEmitDefaultValue(parameter.vendorExtensions);
 
         if (!GENERICHOST.equals(getLibrary()) && !parameter.dataType.endsWith("?") && !parameter.required && (nullReferenceTypesFlag || this.getNullableTypes().contains(parameter.dataType))) {
             parameter.dataType = parameter.dataType + "?";
-        }
-
-        super.postProcessParameter(parameter);
-    }
-
-    /*
-     * See https://msdn.microsoft.com/en-us/library/yd1hzczs(v=vs.110).aspx for .NET options.
-     */
-    public void postProcessPattern(String pattern, Map<String, Object> vendorExtensions) {
-        if (pattern != null) {
-            // check if the pattern has any modifiers
-            // gmiyuvsd - ecma modifiers
-            // l - legacy modifier provided by this library, provides a way to opt out of culture invariant
-            // nx - c# modifiers https://learn.microsoft.com/en-us/dotnet/standard/base-types/regular-expression-options
-            Pattern hasModifiers = Pattern.compile(".*/[gmiyuvsdlnx]+$");
-
-            int end = hasModifiers.matcher(pattern).find()
-                ? pattern.lastIndexOf('/')
-                : pattern.length() - 1;
-
-            int start = pattern.startsWith("/")
-                ? 1
-                : 0;
-
-            Map<Character, String> optionsMap = new HashMap();
-            optionsMap.put('i', "IgnoreCase");
-            optionsMap.put('m', "Multiline");
-            optionsMap.put('s', "SingleLine");
-            optionsMap.put('n', "ExplicitCapture");
-            optionsMap.put('x', "IgnorePatternWhitespace");
-
-            List<String> modifiers = new ArrayList<String>();
-            modifiers.add("CultureInvariant");
-
-            for (char c : pattern.substring(end).toCharArray()) {
-                if (optionsMap.containsKey(c)) {
-                    modifiers.add(optionsMap.get(c));
-                } else if (c == 'l'){
-                    modifiers.remove("CultureInvariant");
-                } else {
-                    vendorExtensions.put("x-modifier-" + c, c);
-                }
-            }
-
-            String regex = pattern.substring(start, end).replace("'", "\'").replace("\"", "\"\"");
-            vendorExtensions.put("x-regex", regex);
-            vendorExtensions.put("x-modifiers", modifiers);
         }
     }
